feat: only enable csr in pro scaffold
