+++ scaffolds/ice3-antd-pro/ice.config.mts
@@ -5,13 +5,11 @@ import auth from '@ice/plugin-auth';
 
 // The project config, see https://v3.ice.work/docs/guide/basic/config
 export default defineConfig({
+  ssr: false,
+  ssg: false,
   plugins: [
     request(),
     store(),
     auth(),
   ],
-  server: {
-    bundle: true,
-    format: 'cjs',
-  },
 });
